LayoutTests/http/tests/site-isolation/history/add-iframe-then-go-back-expected.txt

Verifies that back navigation restores the correct frame tree even when an iframe was dynamically added before navigating away.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS iframeAUrl.includes('id=a&navigated=true') is true
PASS successfullyParsed is true

TEST COMPLETE
